poller: Adapt NIPA for the RISC-V tree
Hack to make the pw_poller.py usable for the RISC-V kernel tree.

The RISC-V does not require the patches to "tree marked", stating what
tree the series/patch is aimed for.

There are two RISC-V trees -- for-next and fixes. Fuzzy guessing
determines to which tree the patch/series is aimed. Now, this is way
too simplistic, if there's a Fixes: tag, we assume the fixes tree.

The tests has been duplicated. The test_rv directory contain the
RISC-V tests, where tests that did not make sense for RISC-V were
removed.

Please refer to the README for more details.

Signed-off-by: Björn Töpel <bjorn@rivosinc.com>