rv: build_warn: try to fix build error/warning reporting
Currently we build the tree 3 times. The first of these builds doesn't
really seem to do anything other than populate the compiler cache & is
build using HEAD.
The second build is of HEAD~1 & is how we determine the number of
"incumbent" issues with the tree.
The third build is of HEAD again, and it is used to figure out which new
issues have been added to the tree.

This however does not actually work as intended. Tuxmake, unlike running
make it self, outputs *its* errors on stderr and outputs warnings/errors
from the build it is running on stdout. The current logic expects
content on stderr & therefore only ever parses the tuxmake post build
report, resulting in zero errors/warnings located - eg:
https://patchwork.kernel.org/project/linux-riscv/patch/20221223221332.4127602-2-heiko@sntech.de/

With this patch applied, I see 2054 warnings/errors in the tree both
before & after.

Perhaps a future optimisation could squash this into two builds, where
the first determines the current state of the tree & the second the
"incumbent" state.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>