rv: build_warn: drop the extra build
This test inherited a 3rd build of the tree from NIPA, but this appears
to serve no real purpose. Currently the order is:
- the tree with the patch
- the tree before the patch
- the tree with the patch
As things stand, run 1 is used to determine if the comparison is worth
doing, IOW if the tree even compiles with this patch. Run 2 is used to
build the tree and establish a baseline number of warnings. Run 3 is
then used to figure out if anything has been added.

Re-order things so that we build the patch & then the tree prior to the
patch. The first run now does what run 1 & 3 used to do - so there
should be no functional difference, but we save an allmodconfig run per
patch.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>