rv: dtb_warn: fix error reporting
I kinda blindly copied from the build_warn test for the dtb one, and the
broken-ness carried over there.
For this test, consider anything on stderr to be an error (but remove
the "From schema" duplicate bits). That should mean that most schema
errors are counted as 1 & other errors/warnings *should* be oneliners.
Hopefully. If they aren't, can deal with them if anyone actually cares.

While at it, drop the extra build.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>