rv: allmodconfig: fix warning breakdown
Currently we diff the stderr to produce a list of what errors/warnings a
patch added. That may be fine if the build order is always the same, but
unfortunately that's not the case - we get something like the link
below.

That's useless for both contributors and those passing the results on to
them. Copy from the per file breakdown code (which I think is not doing
anything useful anyway at the moment) but drop the sed that strips the
errors out. This should produce a diff containing only the differences
in errors & not a huge amount of line-noise.

Link: https://gist.github.com/conor-pwbot/78a79a42c3ac541a957952d6a8743b60
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>