rv: allmodconfig: restore the baseline build
So, turns out we do need the baseline build after all!

Previously, we were doing ccached but not incremental builds so the
whole tree was build each time we called tuxmake. That made the baseline
build redundant, as we would encounter every file and thus their errors
each time.

Now that we are using incremental builds the first build encounters all
files (and therefore all errors) and the second one only encounters the
files touched by the patch. This causes NIPA to report things like:
"Errors and warnings before: 0 this patch: 2054"

Add back the baseline build so that both of the builds used to collect
errors build the same files.

Add in a missing deletion of the builddir while we're at it.

Link: https://patchwork.kernel.org/project/linux-riscv/patch/20230113144552.138081-3-bjorn@kernel.org/
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>